Move extractCodeBlocks REPL.Command -> TypeCheck.AST.
This function is useful to determine when to add implicit :exhaust commands to
property docstrings, but was too high in the Cryptol.* module dependency
hierarchy.
